feat(frontend): Improve how AXE is imported
AXE is an accessibility tool that should prove useful during development.
However, it was being packaged up with the rest of the app as it was
being imported at the top of the file, therefore bloating the size of the
deployed package. This code only imports the package if we are in `dev`
move, so is only used during development.