feat(frontend): Add queryClient privider to test-wrapper, create utility for queryClient
The tests were failing because the test wrapper did now have a queryClient provider
like the app does, so it was added.

Also, this would have meant duplicating the function that instantiates the queryClient
within the provider, which could lead to unexpected side effects if the one in the
app is updated but the one in the tests is forgotten.

To get around this, I created a utility function that just instantiates a queryClient
object with the config and returns it, so that it can be used by both the app and tests.