fix(tests): Make custom renderHooks function more generic
This renderHook custom function will take any hook we can think of.
So, it has to be configurable, meaning we need to configure it using
typescript generics, instead of the hard-coded hook result type. This
means that when using the renderer, each usage passes in the type of
the response and the params passed into it.